requirements.txt
scrapy
chromadb 
sentence-transformers 
pandas
langchain
langchain-groq
pydantic
python-dotenv
fastapi
uvicorn
streamlit
openpyxl